backend/deckShuffle/src/main/python/requirements.txt
torch==2.7.1
scikit-learn==1.7.0
psycopg2-binary
redis